refactor: 迁移领域数据类至 lib/models 以修正分层依赖方向
