+++ lib/features/dashboard/models/dashboard_model.dart
@@ -1,8 +1,10 @@
 import 'package:onetj/services/tongji.dart';
 
 import 'package:onetj/app/di/dependencies.dart';
+import 'package:onetj/models/student_info_data.dart';
 import 'package:onetj/repo/student_info_repository.dart';
-import 'package:onetj/repo/school_calendar_repository.dart';
+import 'package:onetj/models/school_calendar_data.dart';
+import 'package:onetj/models/course_schedule_data.dart';
 import 'package:onetj/repo/course_schedule_repository.dart';
 import 'package:onetj/models/timetable_index.dart';
 import 'package:onetj/services/term_key_resolver.dart';
