+++ lib/features/home/models/home_model.dart
@@ -1,8 +1,8 @@
 import 'package:onetj/services/tongji.dart';
 
-import 'package:onetj/repo/student_info_repository.dart';
-import 'package:onetj/repo/school_calendar_repository.dart';
-import 'package:onetj/repo/course_schedule_repository.dart';
+import 'package:onetj/models/student_info_data.dart';
+import 'package:onetj/models/school_calendar_data.dart';
+import 'package:onetj/models/course_schedule_data.dart';
 
 class HomeModel {
   HomeModel({TongjiApi? api}) : _api = api ?? TongjiApi();
