+++ lib/features/home/view_models/home_view_model.dart
@@ -4,8 +4,10 @@ import 'package:onetj/app/di/dependencies.dart';
 import 'package:onetj/models/base_model.dart';
 import 'package:onetj/features/home/models/home_model.dart';
 import 'package:onetj/models/timetable_index.dart';
-import 'package:onetj/repo/student_info_repository.dart';
+import 'package:onetj/models/student_info_data.dart';
+import 'package:onetj/models/school_calendar_data.dart';
 import 'package:onetj/repo/school_calendar_repository.dart';
+import 'package:onetj/models/course_schedule_data.dart';
 import 'package:onetj/repo/course_schedule_repository.dart';
 import 'package:onetj/services/timetable_index_builder.dart';
