+++ lib/features/settings/models/settings_model.dart
@@ -2,7 +2,7 @@ import 'package:onetj/app/exception/app_exception.dart';
 import 'package:onetj/models/settings_defaults.dart';
 import 'package:onetj/models/settings_validation.dart' as settings_validation;
 import 'package:onetj/models/time_period_range.dart';
-import 'package:onetj/repo/settings_repository.dart';
+import 'package:onetj/models/settings_data.dart';
 
 class SettingsModel {
   static void validateSettings(SettingsData settings) {
