+++ lib/features/timetable/models/timetable_model.dart
@@ -1,7 +1,9 @@
 import 'package:onetj/services/tongji.dart';
 
 import 'package:onetj/app/di/dependencies.dart';
+import 'package:onetj/models/course_schedule_data.dart';
 import 'package:onetj/repo/course_schedule_repository.dart';
+import 'package:onetj/models/school_calendar_data.dart';
 import 'package:onetj/repo/school_calendar_repository.dart';
 import 'package:onetj/services/timetable_index_builder.dart';
 import 'package:onetj/services/term_key_resolver.dart';
