+++ lib/models/timetable_index.dart
@@ -1,4 +1,4 @@
-import 'package:onetj/repo/course_schedule_repository.dart';
+import 'package:onetj/models/course_schedule_data.dart';
 
 /// 课程条目
 ///
