+++ lib/services/term_key_resolver.dart
@@ -1,5 +1,6 @@
 import 'package:onetj/app/di/dependencies.dart';
 import 'package:onetj/repo/course_schedule_repository.dart';
+import 'package:onetj/models/school_calendar_data.dart';
 import 'package:onetj/repo/school_calendar_repository.dart';
 
 class TermKeyResolver {
