+++ test/features/timetable/course_detail_field_mapper_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:onetj/features/timetable/models/course_detail_field_mapper.dart';
-import 'package:onetj/repo/course_schedule_repository.dart';
+import 'package:onetj/models/course_schedule_data.dart';
 import 'package:onetj/services/timetable_index_builder.dart';
 
 void main() {
