+++ test/repo/course_schedule_repository_test.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
+import 'package:onetj/models/course_schedule_data.dart';
 import 'package:onetj/repo/course_schedule_repository.dart';
 
 void main() {
