+++ test/repo/student_info_repository_test.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
+import 'package:onetj/models/student_info_data.dart';
 import 'package:onetj/repo/student_info_repository.dart';
 
 StudentInfoData buildStudentInfo({String deptName = 'Computer Science'}) {
