+++ test/repo/undergraduate_score_repository_test.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
+import 'package:onetj/models/undergraduate_score_data.dart';
 import 'package:onetj/repo/undergraduate_score_repository.dart';
 
 UndergraduateScoreData buildUndergraduateScore({String? totalGradePoint}) {
