+++ test/services/app_update_service_test.dart
@@ -6,6 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 import 'package:onetj/app/exception/app_exception.dart';
 import 'package:onetj/models/app_update_info.dart';
+import 'package:onetj/models/app_update_state_data.dart';
 import 'package:onetj/repo/app_update_state_repository.dart';
 import 'package:onetj/services/app_update_api.dart';
 import 'package:onetj/services/app_update_service.dart';
